Add localizations
Localized strings are defined in resource files (*.resx) and is accessed using an
injected `IStringLocalizer<T>` service.

.NET resource file naming guidelines (links below), should be followed when adding a resource file.

In this sample, `ResourcesPath` options are not configured. Thus, allowing resource files to be
located in the same directory as the targeted class or Razor components for easier management.

For a shared resource that does not have a target class or component, a marker class should be created.
For example, `Common/String.cs` is created as a marker for `Common/Strings.resx` so the localized strings
can be accessed using an injected `IStringLocalizer<Common.Strings>` service.

To organize all resources in a single directory, e.g: in `/Resources/` folder, `ResourcesPath` options
can be set to the directory name, "Resources". Resource files in this directory should be named with
the resource target's full type name, OR, organized in folders, mirroring the target's. For example,
`/Resources/Components.Main.Pages.HomePage.resx` or `/Resources/Components/Main/Pages/HomePage.resx`
can be added for `/Components/Main/Pages/HomePage.razor` component. Similarly, `/Resources/Common.Strings.resx`
or `/Resources/Common/Strings.resx` can be added with a marker class of `/Common/Strings.cs`

See:
 - <https://learn.microsoft.com/en-us/aspnet/core/fundamentals/localization/provide-resources?view=aspnetcore-8.0#resource-file-naming>
 - <https://learn.microsoft.com/en-us/dotnet/standard/design-guidelines/naming-resources>